Create devgroup-role.yaml
creates an IAM DevGroup and assigns it a policy with permissions necessary for the development team. The permissions include:

EC2: Allows describing, starting, and stopping EC2 instances.
RDS: Allows describing RDS DB instances and clusters.
This template is designed to help the development team get the required access to manage EC2 instances and RDS databases without the need for system administrators to configure permissions manually.